drop kserveTarget from ServiceConfig, explainers endpoints package refactoring
