+++ explainability-service/src/main/java/org/kie/trustyai/service/endpoints/explainers/local/CounterfactualEndpoint.java
@@ -48,10 +48,10 @@ import org.kie.trustyai.explainability.utils.DataUtils;
 import org.kie.trustyai.service.config.ServiceConfig;
 import org.kie.trustyai.service.data.DataSource;
 import org.kie.trustyai.service.data.exceptions.DataframeCreateException;
-import org.kie.trustyai.service.payloads.explainability.BaseExplanationResponse;
-import org.kie.trustyai.service.payloads.explainability.CounterfactualExplanationRequest;
-import org.kie.trustyai.service.payloads.explainability.CounterfactualExplanationResponse;
-import org.kie.trustyai.service.payloads.explainability.LocalExplanationRequest;
+import org.kie.trustyai.service.payloads.explainers.BaseExplanationResponse;
+import org.kie.trustyai.service.payloads.explainers.CounterfactualExplanationRequest;
+import org.kie.trustyai.service.payloads.explainers.CounterfactualExplanationResponse;
+import org.kie.trustyai.service.payloads.explainers.LocalExplanationRequest;
 
 @Tag(name = "Counterfactual Explainer Endpoint")
 @Path("/explainers/local/cf")
