+++ explainability-service/src/main/java/org/kie/trustyai/service/endpoints/explainers/local/SHAPEndpoint.java
@@ -36,9 +36,9 @@ import org.kie.trustyai.explainability.model.PredictionInput;
 import org.kie.trustyai.explainability.model.PredictionProvider;
 import org.kie.trustyai.service.config.ServiceConfig;
 import org.kie.trustyai.service.data.DataSource;
-import org.kie.trustyai.service.payloads.explainability.BaseExplanationResponse;
-import org.kie.trustyai.service.payloads.explainability.LocalExplanationRequest;
-import org.kie.trustyai.service.payloads.explainability.SaliencyExplanationResponse;
+import org.kie.trustyai.service.payloads.explainers.BaseExplanationResponse;
+import org.kie.trustyai.service.payloads.explainers.LocalExplanationRequest;
+import org.kie.trustyai.service.payloads.explainers.SaliencyExplanationResponse;
 
 @Tag(name = "SHAP Explainer Endpoint")
 @Path("/explainers/local/shap")
