+++ explainability-service/src/main/java/org/kie/trustyai/service/payloads/explainers/LocalExplanationRequest.java
@@ -13,23 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.trustyai.service.payloads.explainability;
+package org.kie.trustyai.service.payloads.explainers;
 
 public class LocalExplanationRequest {
 
-    private String modelId;
-
     private String predictionId;
 
-    private String modelVersion;
-
-    public String getModelId() {
-        return modelId;
-    }
-
-    public void setModelId(String modelId) {
-        this.modelId = modelId;
-    }
+    private ModelConfig modelConfig;
 
     public String getPredictionId() {
         return predictionId;
@@ -39,12 +29,11 @@ public class LocalExplanationRequest {
         this.predictionId = predictionId;
     }
 
-    public String getModelVersion() {
-        return modelVersion;
+    public ModelConfig getModelConfig() {
+        return modelConfig;
     }
 
-    public void setModelVersion(String modelVersion) {
-        this.modelVersion = modelVersion;
+    public void setModelConfig(ModelConfig modelConfig) {
+        this.modelConfig = modelConfig;
     }
-
 }
