+++ explainability-service/src/test/java/org/kie/trustyai/service/endpoints/explainers/local/CounterfactualEndpointTest.java
@@ -15,7 +15,8 @@ import org.kie.trustyai.explainability.model.PredictionInput;
 import org.kie.trustyai.service.endpoints.explainers.ExplainersEndpointTestProfile;
 import org.kie.trustyai.service.mocks.MockDatasource;
 import org.kie.trustyai.service.mocks.MockMemoryStorage;
-import org.kie.trustyai.service.payloads.explainability.CounterfactualExplanationRequest;
+import org.kie.trustyai.service.payloads.explainers.CounterfactualExplanationRequest;
+import org.kie.trustyai.service.payloads.explainers.ModelConfig;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 
@@ -63,7 +64,7 @@ class CounterfactualEndpointTest {
         List<PredictionInput> predictionInputs = dataframe.asPredictionInputs();
         String id = String.valueOf(predictionInputs.get(0).hashCode());
         final CounterfactualExplanationRequest payload = new CounterfactualExplanationRequest();
-        payload.setModelId(MODEL_ID);
+        payload.setModelConfig(new ModelConfig("", MODEL_ID, ""));
         payload.setPredictionId(id);
         Map<String, String> map = new HashMap<>();
         map.put("income", "2");
