+++ explainability-service/src/test/java/org/kie/trustyai/service/mocks/MockPVCServiceConfig.java
@@ -1,6 +1,5 @@
 package org.kie.trustyai.service.mocks;
 
-import java.util.Optional;
 import java.util.OptionalInt;
 
 import org.kie.trustyai.service.config.ServiceConfig;
@@ -8,14 +7,10 @@ import org.kie.trustyai.service.data.storage.DataFormat;
 import org.kie.trustyai.service.data.storage.StorageFormat;
 
 public class MockPVCServiceConfig implements ServiceConfig {
+
     @Override
     public OptionalInt batchSize() {
         return OptionalInt.of(5000);
-    }
-
-    @Override
-    public Optional<String> kserveTarget() {
-        return Optional.empty();
     }
 
     @Override
